+++ frontend/app-development/router/routes.tsx
@@ -1,7 +1,7 @@
 import { SubApp } from '../../packages/ux-editor/src/SubApp';
 import { Administration } from '../features/administration/components/Administration';
 import { LegacyAdministration } from '../features/administration/components/LegacyAdministration';
-import { Contact } from '../features/Contact';
+import { Contact } from '../features/administration/components/Contact';
 import { TextEditor } from '../features/textEditor/TextEditor';
 import DataModellingContainer from '../features/dataModelling/containers/DataModellingContainer';
 import { DeployPage } from '../features/appPublish/pages/deployPage';
